feat: implementação do typeDI feita.
