+++ src/resolvers/address-reolvers.ts
@@ -4,15 +4,15 @@ import { AddressModel } from '../model/address-model';
 import { CreateAddressInput } from './input-validation/address-input-validation';
 import { IMyContext } from '../types/type-context';
 import jwtUtil from '../utils/jwt-util';
-import { Inject, Service } from 'typedi';
+import Container, { Service } from 'typedi';
 
 @Service()
 @Resolver()
 export class AddressResolvers {
   private AddressService: AddressService;
 
-  constructor(@Inject() AddressService: AddressService) {
-    this.AddressService = AddressService;
+  constructor() {
+    this.AddressService = Container.get(AddressService);
   }
 
   @Mutation(() => AddressModel)
